Split validation versioned namespaces
