+++ libmamba/include/mamba/core/validate.hpp
@@ -113,7 +113,7 @@ namespace mamba::validation
 
     protected:
 
-        SpecBase(const std::string& spec_version);
+        SpecBase(std::string spec_version);
         SpecBase() = delete;
 
         std::string get_json_value(const nlohmann::json& j) const;
@@ -307,269 +307,5 @@ namespace mamba::validation
 
         std::unique_ptr<RootRole> get_root_role(const TimeRef& time_reference);
     };
-
-
-    namespace v1
-    {
-        /**
-         * TUF v1 specific implementation.
-         */
-        class SpecImpl final : public SpecBase
-        {
-        public:
-
-            SpecImpl(const std::string& sv = "1.0.17");
-
-            std::string json_key() const override;
-            std::string expiration_json_key() const override;
-
-            std::set<RoleSignature> signatures(const nlohmann::json& j) const override;
-        };
-
-
-        /**
-         * 'root' role implementation.
-         * TUF v1.0.17 §2.1.1
-         * https://theupdateframework.github.io/specification/latest/#root
-         */
-        class RootImpl final : public RootRole
-        {
-        public:
-
-            RootImpl(const fs::u8path& p);
-            RootImpl(const nlohmann::json& j);
-
-            RoleFullKeys self_keys() const override;
-
-            std::unique_ptr<RepoIndexChecker> build_index_checker(
-                Context& context,
-                const TimeRef& time_reference,
-                const std::string& url,
-                const fs::u8path& cache_path
-            ) const override;
-
-            friend void to_json(nlohmann::json& j, const RootImpl& r);
-            friend void from_json(const nlohmann::json& j, RootImpl& r);
-
-        private:
-
-            RootImpl() = delete;
-
-            void load_from_json(const nlohmann::json& j);
-
-            std::unique_ptr<RootRole> create_update(const nlohmann::json& j) override;
-
-            std::set<std::string> mandatory_defined_roles() const override;
-            std::set<std::string> optionally_defined_roles() const override;
-
-            void
-            set_defined_roles(std::map<std::string, Key> keys, std::map<std::string, RoleKeys> roles);
-        };
-    }
-
-    namespace v06
-    {
-        /**
-         * ``conda-content-trust`` v0.6.0 specific implementation.
-         * This is a variation of TUF specification.
-         */
-        class SpecImpl final : public SpecBase
-        {
-        public:
-
-            SpecImpl(const std::string& sv = "0.6.0");
-
-            std::string json_key() const override;
-            std::string expiration_json_key() const override;
-
-            std::set<RoleSignature> signatures(const nlohmann::json& j) const override;
-
-            std::string canonicalize(const nlohmann::json& j) const override;
-            bool upgradable() const override;
-        };
-
-        class V06RoleBaseExtension
-        {
-        public:
-
-            void set_timestamp(const std::string& ts);
-
-            std::string timestamp() const;
-
-        protected:
-
-            std::string m_timestamp;
-
-            void check_timestamp_format() const;
-        };
-
-
-        // Forward declaration of KeyMgrRole.
-        class KeyMgrRole;
-
-        /**
-         * 'root' role implementation.
-         */
-        class RootImpl final
-            : public RootRole
-            , public V06RoleBaseExtension
-        {
-        public:
-
-            RootImpl(const fs::u8path& p);
-            RootImpl(const nlohmann::json& j);
-            RootImpl(const std::string& json_str);
-
-            /**
-             * Return a ``RepoIndexChecker`` implementation (derived class)
-             * from repository base URL.
-             */
-            std::unique_ptr<RepoIndexChecker> build_index_checker(
-                Context& context,
-                const TimeRef& time_reference,
-                const std::string& url,
-                const fs::u8path& cache_path
-            ) const override;
-
-            RoleFullKeys self_keys() const override;
-
-            nlohmann::json upgraded_signable() const;
-            RoleSignature
-            upgraded_signature(const nlohmann::json& j, const std::string& pk, const std::byte* sk) const;
-
-            KeyMgrRole create_key_mgr(const fs::u8path& p) const;
-            KeyMgrRole create_key_mgr(const nlohmann::json& j) const;
-
-            friend void to_json(nlohmann::json& j, const RootImpl& r);
-            friend void from_json(const nlohmann::json& j, RootImpl& r);
-
-        private:
-
-            RootImpl() = delete;
-
-            void load_from_json(const nlohmann::json& j);
-
-            std::unique_ptr<RootRole> create_update(const nlohmann::json& j) override;
-
-            std::set<std::string> mandatory_defined_roles() const override;
-            std::set<std::string> optionally_defined_roles() const override;
-
-            void set_defined_roles(std::map<std::string, RolePubKeys> keys);
-        };
-
-
-        // Forward declaration of KeyMgrRole.
-        class PkgMgrRole;
-
-        /**
-         * 'key_mgr' role implementation.
-         */
-        class KeyMgrRole final
-            : public RoleBase
-            , public V06RoleBaseExtension
-        {
-        public:
-
-            KeyMgrRole(const fs::u8path& p, const RoleFullKeys& keys, const std::shared_ptr<SpecBase> spec);
-            KeyMgrRole(
-                const nlohmann::json& j,
-                const RoleFullKeys& keys,
-                const std::shared_ptr<SpecBase> spec
-            );
-            KeyMgrRole(
-                const std::string& json_str,
-                const RoleFullKeys& keys,
-                const std::shared_ptr<SpecBase> spec
-            );
-
-            // std::set<std::string> roles() const override;
-            RoleFullKeys self_keys() const override;
-
-            PkgMgrRole create_pkg_mgr(const fs::u8path& p) const;
-            PkgMgrRole create_pkg_mgr(const nlohmann::json& j) const;
-
-            /**
-             * Return a ``RepoIndexChecker`` implementation (derived class)
-             * from repository base URL.
-             */
-            std::unique_ptr<RepoIndexChecker> build_index_checker(
-                Context& context,
-                const TimeRef& time_reference,
-                const std::string& url,
-                const fs::u8path& cache_path
-            ) const;
-
-            friend void to_json(nlohmann::json& j, const KeyMgrRole& r);
-            friend void from_json(const nlohmann::json& j, KeyMgrRole& r);
-
-        private:
-
-            KeyMgrRole() = delete;
-
-            void load_from_json(const nlohmann::json& j);
-
-            RoleFullKeys m_keys;
-            std::map<std::string, RolePubKeys> m_delegations;
-
-            std::set<std::string> mandatory_defined_roles() const override;
-            std::set<std::string> optionally_defined_roles() const override;
-
-            void set_defined_roles(std::map<std::string, RolePubKeys> keys);
-        };
-
-
-        /**
-         * 'pkg_mgr' role implementation.
-         * This role inherits from ``RepoIndexChecker`` and
-         * will be used by ``RepoChecker`` to perform the
-         * repository index verification.
-         */
-        class PkgMgrRole final
-            : public RoleBase
-            , public V06RoleBaseExtension
-            , public RepoIndexChecker
-        {
-        public:
-
-            PkgMgrRole(const RoleFullKeys& keys, const std::shared_ptr<SpecBase> spec);
-            PkgMgrRole(const fs::u8path& p, const RoleFullKeys& keys, const std::shared_ptr<SpecBase> spec);
-            PkgMgrRole(
-                const nlohmann::json& j,
-                const RoleFullKeys& keys,
-                const std::shared_ptr<SpecBase> spec
-            );
-            PkgMgrRole(
-                const std::string& json_str,
-                const RoleFullKeys& keys,
-                const std::shared_ptr<SpecBase> spec
-            );
-
-            void verify_index(const fs::u8path& p) const override;
-            void verify_index(const nlohmann::json& j) const override;
-            void verify_package(const nlohmann::json& signed_data, const nlohmann::json& signatures)
-                const override;
-
-            friend void to_json(nlohmann::json& j, const PkgMgrRole& r);
-            friend void from_json(const nlohmann::json& j, PkgMgrRole& r);
-
-        private:
-
-            PkgMgrRole() = delete;
-
-            void load_from_json(const nlohmann::json& j);
-
-            RoleFullKeys self_keys() const override;
-            std::set<RoleSignature> pkg_signatures(const nlohmann::json& j) const;
-            void
-            check_pkg_signatures(const nlohmann::json& signed_data, const nlohmann::json& signatures) const;
-
-            void set_defined_roles(std::map<std::string, RolePubKeys> keys);
-
-            RoleFullKeys m_keys;
-
-            friend class KeyMgrRole;
-        };
-    }
-}  // namespace validate
-
-#endif  // MAMBA_VALIDATE_HPP
+}
+#endif
